Improve extension activation tests
